Final Project v3.0 - Final Version Monday Night
Solid prototype but still needs a lot of work to make it more functional.  I had to move from NYC the last few weeks so I have not had time to spend on making it "more pretty" with modal and other features, but I do think I have a solid start.  I look forward to making it an incredible site over the next 6 months and launch live one day as a true e-commerce site.